frontend done

+++ Backend/src/utils/cloudinary.js
@@ -20,7 +20,9 @@ import fs from 'fs';
         return response;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+        if (localFilePath && fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+        }
         return null;
     }
 }
